requirements.txt
google-adk
python-dotenv
# Add litellm if using non-Google models via LiteLlm